requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
selenium==4.15.2
webdriver-manager==4.0.1
pdfkit==1.0.0
reportlab==4.0.9
python-dateutil==2.8.2